make locator async
